src/requirements.txt
Flask
flask-cors
google-generativeai
gunicorn
numpy
opencv-python
Pillow
